+++ ogcapi-draft/ogcapi-transactional/src/main/java/de/ii/ogcapi/crud/app/HeaderLocationCrud.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.transactional.app;
+package de.ii.ogcapi.crud.app;
 
 import com.github.azahnen.dagger.annotations.AutoBind;
 import de.ii.ogcapi.foundation.domain.ApiExtensionCache;
@@ -21,13 +21,13 @@ import javax.inject.Singleton;
 
 @Singleton
 @AutoBind
-public class HeaderLocationTransactional extends ApiExtensionCache implements ApiHeader {
+public class HeaderLocationCrud extends ApiExtensionCache implements ApiHeader {
 
   private final Schema<?> schema = new StringSchema().format("uri");
   private final SchemaValidator schemaValidator;
 
   @Inject
-  HeaderLocationTransactional(SchemaValidator schemaValidator) {
+  HeaderLocationCrud(SchemaValidator schemaValidator) {
     this.schemaValidator = schemaValidator;
   }
