+++ ogcapi-draft/ogcapi-transactional/src/test/groovy/de/ii/ogcapi/crud/TransactionalRESTApiSpec.groovy
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.ogcapi.transactional
+package de.ii.ogcapi.crud
 
 
 import groovyx.net.http.*
